Simplify & cleanup gravatar fallback; dont delete image if none exists

+++ app/User.php
@@ -66,7 +66,7 @@ class User extends Authenticatable
     public function getProfilePictureHiresAttribute()
     {
         if (! $this->profile_picture) {
-            return Gravatar::src($this->email);
+            return Gravatar::src($this->email, 500);
         }
 
         return asset('/storage/' . self::PROFILE_PICTURE_HIRES_PATH . $this->profile_picture);
@@ -78,14 +78,16 @@ class User extends Authenticatable
 
         // Cascade deletes
         static::deleting(function ($user) {
-             $user->talks()->delete();
-             // $user->conferences()->delete(); // Not sure if we want to do this.
-             $user->bios()->delete();
+            $user->talks()->delete();
+            // $user->conferences()->delete(); // Not sure if we want to do this.
+            $user->bios()->delete();
 
-             Storage::delete(User::PROFILE_PICTURE_THUMB_PATH . $user->profile_picture);
-             Storage::delete(User::PROFILE_PICTURE_HIRES_PATH . $user->profile_picture);
+            if ($user->profile_picture && strpos($user->profile_picture, '/') === false) {
+                Storage::delete(User::PROFILE_PICTURE_THUMB_PATH . $user->profile_picture);
+                Storage::delete(User::PROFILE_PICTURE_HIRES_PATH . $user->profile_picture);
+            }
 
-             \DB::table('favorites')->where('user_id', $user->id)->delete();
+            \DB::table('favorites')->where('user_id', $user->id)->delete();
         });
     }
 }
